+++ examples/react-renderer/src/Autocomplete.tsx
@@ -52,6 +52,7 @@ export function Autocomplete(
         getSources() {
           return [
             {
+              sourceId: 'products',
               getItems({ query }) {
                 return getAlgoliaHits({
                   searchClient,
@@ -192,7 +193,7 @@ export function Autocomplete(
                                 width="20"
                                 height="20"
                               >
-                                <path d="M18.984 6.984h2.016v6h-15.188l3.609 3.609-1.406 1.406-6-6 6-6 1.406 1.406-3.609 3.609h13.172v-4.031z"></path>
+                                <path d="M18.984 6.984h2.016v6h-15.188l3.609 3.609-1.406 1.406-6-6 6-6 1.406 1.406-3.609 3.609h13.172v-4.031z" />
                               </svg>
                             </button>
                           </li>