+++ packages/autocomplete-js/src/autocomplete.ts
@@ -36,6 +36,12 @@ export function autocomplete<TItem extends BaseItem>(
     AutocompleteOptions<TItem>['onStateChange']
   >(undefined);
   const props = reactive(() => getDefaultOptions(optionsRef.current));
+  const isDetached = reactive(
+    () =>
+      props.value.core.environment.matchMedia(
+        props.value.renderer.detachedMediaQuery
+      ).matches
+  );
   const autocomplete = reactive(() =>
     createAutocomplete<TItem>({
       ...props.value.core,
@@ -47,8 +53,8 @@ export function autocomplete<TItem extends BaseItem>(
         onStateChangeRef.current?.(options as any);
         props.value.core.onStateChange?.(options as any);
       },
-      shouldPanelShow:
-        optionsRef.current.shouldPanelShow ||
+      shouldPanelOpen:
+        optionsRef.current.shouldPanelOpen ||
         (({ state }) => {
           const hasItems = getItemsCount(state) > 0;
 
@@ -76,10 +82,6 @@ export function autocomplete<TItem extends BaseItem>(
     ...props.value.core.initialState,
   });
 
-  const isTouch = reactive(
-    () => window.matchMedia(props.value.renderer.touchMediaQuery).matches
-  );
-
   const propGetters: AutocompletePropGetters<TItem> = {
     getEnvironmentProps: props.value.renderer.getEnvironmentProps,
     getFormProps: props.value.renderer.getFormProps,
@@ -105,7 +107,7 @@ export function autocomplete<TItem extends BaseItem>(
       autocomplete: autocomplete.value,
       autocompleteScopeApi,
       classNames: props.value.renderer.classNames,
-      isTouch: isTouch.value,
+      isDetached: isDetached.value,
       placeholder: props.value.core.placeholder,
       propGetters,
       state: lastStateRef.current,
@@ -114,7 +116,7 @@ export function autocomplete<TItem extends BaseItem>(
 
   function setPanelPosition() {
     setProperties(dom.value.panel, {
-      style: isTouch.value
+      style: isDetached.value
         ? {}
         : getPanelPlacementStyle({
             panelPlacement: props.value.renderer.panelPlacement,
@@ -136,9 +138,8 @@ export function autocomplete<TItem extends BaseItem>(
       createElement: props.value.renderer.renderer.createElement,
       dom: dom.value,
       Fragment: props.value.renderer.renderer.Fragment,
-      isTouch: isTouch.value,
-      panelContainer: isTouch.value
-        ? dom.value.touchOverlay
+      panelContainer: isDetached.value
+        ? dom.value.detachedOverlay
         : props.value.renderer.panelContainer,
       propGetters,
       state: lastStateRef.current,
@@ -161,11 +162,11 @@ export function autocomplete<TItem extends BaseItem>(
       inputElement: dom.value.input,
     });
 
-    setProperties(window as any, environmentProps);
+    setProperties(props.value.core.environment as any, environmentProps);
 
     return () => {
       setProperties(
-        window as any,
+        props.value.core.environment as any,
         Object.keys(environmentProps).reduce((acc, key) => {
           return {
             ...acc,
@@ -177,15 +178,15 @@ export function autocomplete<TItem extends BaseItem>(
   });
 
   runEffect(() => {
-    const panelContainerElement = isTouch.value
+    const panelContainerElement = isDetached.value
       ? document.body
       : props.value.renderer.panelContainer;
-    const panelElement = isTouch.value
-      ? dom.value.touchOverlay
+    const panelElement = isDetached.value
+      ? dom.value.detachedOverlay
       : dom.value.panel;
 
-    if (isTouch.value && lastStateRef.current.isOpen) {
-      dom.value.openTouchOverlay();
+    if (isDetached.value && lastStateRef.current.isOpen) {
+      dom.value.openDetachedOverlay();
     }
 
     scheduleRender(lastStateRef.current);
@@ -232,21 +233,21 @@ export function autocomplete<TItem extends BaseItem>(
 
   runEffect(() => {
     const onResize = debounce<Event>(() => {
-      const previousIsTouch = isTouch.value;
-      isTouch.value = window.matchMedia(
-        props.value.renderer.touchMediaQuery
+      const previousisDetached = isDetached.value;
+      isDetached.value = props.value.core.environment.matchMedia(
+        props.value.renderer.detachedMediaQuery
       ).matches;
 
-      if (previousIsTouch !== isTouch.value) {
+      if (previousisDetached !== isDetached.value) {
         update({});
       } else {
         requestAnimationFrame(setPanelPosition);
       }
     }, 20);
-    window.addEventListener('resize', onResize);
+    props.value.core.environment.addEventListener('resize', onResize);
 
     return () => {
-      window.removeEventListener('resize', onResize);
+      props.value.core.environment.removeEventListener('resize', onResize);
     };
   });