+++ packages/autocomplete-js/src/types/AutocompleteOptions.ts
@@ -7,6 +7,7 @@ import {
 import { MaybePromise } from '@algolia/autocomplete-shared';
 
 import { AutocompleteClassNames } from './AutocompleteClassNames';
+import { AutocompletePlugin } from './AutocompletePlugin';
 import { AutocompletePropGetters } from './AutocompletePropGetters';
 import { AutocompleteRender } from './AutocompleteRender';
 import { AutocompleteRenderer } from './AutocompleteRenderer';
@@ -37,12 +38,12 @@ export interface AutocompleteOptions<TItem extends BaseItem>
    */
   panelContainer?: string | HTMLElement;
   /**
-   * The Media Query to turn Autocomplete into a touch experience.
+   * The Media Query to turn Autocomplete into a detached experience.
    *
    * @default "(hover: none) and (pointer: coarse)"
    * @see https://developer.mozilla.org/en-US/docs/Web/CSS/Media_Queries/Using_media_queries
    */
-  touchMediaQuery?: string;
+  detachedMediaQuery?: string;
   getSources?: (
     params: GetSourcesParams<TItem>
   ) => MaybePromise<Array<AutocompleteSource<TItem>>>;
@@ -71,4 +72,5 @@ export interface AutocompleteOptions<TItem extends BaseItem>
    * Custom renderer.
    */
   renderer?: AutocompleteRenderer;
+  plugins?: Array<AutocompletePlugin<TItem, unknown>>;
 }