+++ packages/autocomplete-plugin-algolia-insights/src/types/EventParams.ts
@@ -3,12 +3,12 @@ import { AutocompleteState } from '@algolia/autocomplete-core';
 import {
   ClickedObjectIDsAfterSearchParams,
   ViewedObjectIDsParams,
-} from './InsightsApi';
+} from './AutocompleteInsightsApi';
 
-import { AlgoliaInsightsHit, InsightsApi } from '.';
+import { AlgoliaInsightsHit, AutocompleteInsightsApi } from '.';
 
 export type OnSelectParams = {
-  insights: InsightsApi;
+  insights: AutocompleteInsightsApi;
   insightsEvents: ClickedObjectIDsAfterSearchParams[];
   item: AlgoliaInsightsHit;
   state: AutocompleteState<any>;
@@ -18,7 +18,7 @@ export type OnSelectParams = {
 export type OnActiveParams = OnSelectParams;
 
 export type OnItemsChangeParams = {
-  insights: InsightsApi;
+  insights: AutocompleteInsightsApi;
   insightsEvents: ViewedObjectIDsParams[];
   state: AutocompleteState<any>;
 };