+++ packages/autocomplete-plugin-recent-searches/src/createRecentSearchesPlugin.ts
@@ -18,6 +18,7 @@ export type CreateRecentSearchesPluginParams<
   transformSource?(params: {
     source: AutocompleteSource<TItem>;
     onRemove(id: string): void;
+    onTapAhead(item: TItem): void;
   }): AutocompleteSource<TItem>;
 };
 
@@ -36,10 +37,11 @@ export function createRecentSearchesPlugin<TItem extends RecentSearchesItem>({
       onSelect(({ item, state, source }) => {
         const inputValue = source.getItemInputValue({ item, state });
 
-        if (inputValue) {
+        if (source.sourceId === 'querySuggestionsPlugin' && inputValue) {
           store.add({
-            id: inputValue,
+            objectID: inputValue,
             query: inputValue,
+            category: (item as any).__autocomplete_qsCategory,
           } as any);
         }
       });
@@ -49,12 +51,12 @@ export function createRecentSearchesPlugin<TItem extends RecentSearchesItem>({
 
       if (query) {
         store.add({
-          id: query,
+          objectID: query,
           query,
         } as any);
       }
     },
-    getSources({ query, refresh }) {
+    getSources({ query, setQuery, refresh }) {
       lastItemsRef.current = store.getAll(query);
 
       function onRemove(id: string) {
@@ -62,7 +64,10 @@ export function createRecentSearchesPlugin<TItem extends RecentSearchesItem>({
         refresh();
       }
 
-      const templates = getTemplates<any>({ onRemove });
+      function onTapAhead(item: TItem) {
+        setQuery(item.query);
+        refresh();
+      }
 
       return Promise.resolve(lastItemsRef.current).then((items) => {
         if (items.length === 0) {
@@ -79,9 +84,10 @@ export function createRecentSearchesPlugin<TItem extends RecentSearchesItem>({
               getItems() {
                 return items;
               },
-              templates,
+              templates: getTemplates({ onRemove, onTapAhead }),
             },
             onRemove,
+            onTapAhead,
           }),
         ];
       });