+++ packages/autocomplete-plugin-recent-searches/src/types/RecentSearchesItem.ts
@@ -1,4 +1,5 @@
 export type RecentSearchesItem = {
-  id: string;
+  objectID: string;
   query: string;
+  category?: string;
 };